fex bugs fixed and project code restructured

+++ src/main/java/com/thoughtworks/taskmaster/repositories/TaskRepository.java
@@ -7,7 +7,7 @@ import java.util.List;
 import java.util.Optional;
 
 @Repository
-public interface TaskRepository extends JpaRepository<Task, Integer> {
+public interface TaskRepository extends JpaRepository<Task, Long> {
 
     Optional<List<Task>> findAllByUserId(long userId);
